app/requirements.txt
# Full dependencies (includes AI Vision)
# Use this for development or if you want AI image description

# Core
markitdown>=0.1.0
customtkinter>=5.2.0
Pillow>=10.0.0

# Document support
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.0.0

# PDF image extraction
PyMuPDF>=1.23.0

# AI Vision (optional - adds ~50MB to build)
openai>=1.0.0
google-generativeai>=0.3.0
chardet>=5.0.0
